+++ src/Vision/models/Models.jl
@@ -8,6 +8,7 @@ using BSON
 using Flux
 using Zygote
 using DataDeps
+using InlineTest
 
 
 include("layers.jl")
@@ -16,6 +17,6 @@ include("xresnet.jl")
 include("unet.jl")
 
 
-export xresnet18, xresnet50, UNetDynamic, TabularModel
+export xresnet18, xresnet50, UNetDynamic
 
 end